I've completed the 3rd piece of instructor feedback on my oridinal mailroom.py: "don't forget this report is supposed to be sorted (check requirements)"
I'm not sure I'm going to do the 4th an final piece of feedback, as it's likely still above my skillset. However if you can give me a few tips or point me in the right direction I might give it a shot: "a better alternative to calling prompt() all over the place is to have a while True loop so that it forever loops until program is terminated." Should I put the whole thing in a while loop or just certain parts?

Now moving on to actually starting mailroom2.py